DEF[#79403]: 物联设备新增修改，modbus协议支持 modbus-rtu over tcp 和 modbus-tcp

See merge request iotcenter/IotCenter!1693

+++ iot-center-platform/src/main/java/com/yjh/platform/module/iot/entity/TIotDeviceData.java
@@ -67,7 +67,7 @@ public class TIotDeviceData implements Serializable {
     /**
      * 系数
      */
-    private Integer magnificationCoefficient;
+    private Float magnificationCoefficient;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
